+++ examples/tests/sptrsm_rowmaj_test.hpp
@@ -16,13 +16,13 @@ REGISTER_TEST(sptrsm_rowmaj_test) {
     A->val = new double[A->nnz]{1.1, 2.2, 3.1, 3.3};
 
     double *X = new double[A->n_cols * n_vectors];
-    for (int i = 0; i < A->n_cols * n_vectors; ++i) {
+    for (ULL i = 0; i < A->n_cols * n_vectors; ++i) {
         X[i] = 1.0;
     }
 
     // Initialize RHS
     double *B = new double[A->n_cols * n_vectors];
-    for (int i = 0; i < A->n_cols * n_vectors; ++i) {
+    for (ULL i = 0; i < A->n_cols * n_vectors; ++i) {
         B[i] = 2.0;
     }
