+++ examples/tests/sptrsv_test.hpp
@@ -14,13 +14,13 @@ REGISTER_TEST(sptrsv_test) {
     A->val = new double[A->nnz]{1.1, 2.2, 3.1, 3.3};
 
     double *x = new double[A->n_cols];
-    for (int i = 0; i < A->n_cols; ++i) {
+    for (ULL i = 0; i < A->n_cols; ++i) {
         x[i] = 1.0;
     }
 
     // Initialize RHS
     double *b = new double[A->n_cols];
-    for (int i = 0; i < A->n_cols; ++i) {
+    for (ULL i = 0; i < A->n_cols; ++i) {
         b[i] = 2.0;
     }
